+++ footystats-tools-backend/src/main/java/de/ludwig/footystats/tools/backend/services/prediction/InfluencerPercentDistribution.java
@@ -13,7 +13,7 @@ public class InfluencerPercentDistribution
 {
 	@Getter
 	@Setter
-	private Float predictionPercent;
+	private Integer predictionPercent;
 	@Getter
 	@Setter
 	private Long count;
