+++ footystats-tools-backend/src/main/java/de/ludwig/footystats/tools/backend/services/prediction/influencer/FootyStatsBttsYesPredictionInfluencer.java
@@ -17,7 +17,7 @@ public class FootyStatsBttsYesPredictionInfluencer implements BetResultInfluence
     }
 
     // eslint-disable-next-line class-methods-use-this
-    public Float calculateInfluence(BetPredictionContext ctx) {
+    public Integer calculateInfluence(BetPredictionContext ctx) {
         return ctx.match().getBTTSAverage();
     }
